+++ nextstep-di/src/main/java/nextstep/di/factory/ApplicationContext.java
@@ -1,5 +1,8 @@
 package nextstep.di.factory;
 
+import nextstep.di.scanner.ClasspathBeanScanner;
+import nextstep.di.scanner.ConfigurationBeanScanner;
+
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
 import java.util.Collection;
