+++ nextstep-di/src/main/java/nextstep/di/scanner/BeanScanner.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory;
+package nextstep.di.scanner;
 
 import org.reflections.Reflections;
