+++ nextstep-di/src/test/java/nextstep/di/example/ExampleConfig.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.example;
+package nextstep.di.example;
 
 import nextstep.annotation.Bean;
 import nextstep.annotation.Configuration;
