+++ nextstep-di/src/test/java/nextstep/di/example/IntegrationConfig.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.example;
+package nextstep.di.example;
 
 import nextstep.annotation.Bean;
 import nextstep.annotation.ComponentScan;
@@ -8,7 +8,7 @@ import org.apache.commons.dbcp2.BasicDataSource;
 import javax.sql.DataSource;
 
 @Configuration
-@ComponentScan(basePackages = "nextstep.di.factory.example")
+@ComponentScan(basePackages = "nextstep.di.example")
 public class IntegrationConfig {
     @Bean
     public DataSource dataSource() {
