+++ nextstep-di/src/test/java/nextstep/di/example/JdbcUserRepository.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.example;
+package nextstep.di.example;
 
 import nextstep.annotation.Inject;
 import nextstep.stereotype.Repository;
