+++ nextstep-di/src/test/java/nextstep/di/example/MyJdbcTemplate.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.example;
+package nextstep.di.example;
 
 import javax.sql.DataSource;
