+++ nextstep-di/src/test/java/nextstep/di/example/MyQnaService.java
@@ -1,4 +1,4 @@
-package nextstep.di.factory.example;
+package nextstep.di.example;
 
 public interface MyQnaService {
     UserRepository getUserRepository();
