+++ nextstep-di/src/test/java/nextstep/di/scanner/ClasspathBeanScannerTest.java
@@ -1,9 +1,9 @@
-package nextstep.di.factory;
+package nextstep.di.scanner;
 
-import nextstep.di.factory.example.JdbcQuestionRepository;
-import nextstep.di.factory.example.JdbcUserRepository;
-import nextstep.di.factory.example.MyQnaServiceImpl;
-import nextstep.di.factory.example.QnaController;
+import nextstep.di.example.JdbcQuestionRepository;
+import nextstep.di.example.JdbcUserRepository;
+import nextstep.di.example.MyQnaServiceImpl;
+import nextstep.di.example.QnaController;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -16,7 +16,7 @@ class ClasspathBeanScannerTest {
     @Test
     @DisplayName("@Controller, @Service, @Repository 스캔 확인")
     void scan_Test() {
-        BeanScanner beanScanner = new ClasspathBeanScanner("nextstep.di.factory.example");
+        BeanScanner beanScanner = new ClasspathBeanScanner("nextstep.di.example");
         Set<Class<?>> actual = Set.of(QnaController.class, JdbcUserRepository.class, JdbcQuestionRepository.class, MyQnaServiceImpl.class);
         Set<Class<?>> expected = beanScanner.getClassTypes();
         assertThat(expected).containsAll(actual);
